Update rolling_reboot.nhc to incorporate bug fixes
I am adding some changes that we made locally at our site after we noticed that the node State and Reason was being truncated after the first word.  Additionally now we use regex to grep for certain words or phrases from the Reason line instead of just searching for a specific word, and expanded the State conditions that would qualify a node to have remediation blocks trigger. Additionally we show how we introduce some messaging commands in the action blocks so that operations can be aware of actions being taken by the script.